% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NegtiveMargin.R
\name{CG_Nmargin}
\alias{CG_Nmargin}
\title{CG_Nmargin}
\usage{
CG_Nmargin(nsid, nspwd)
}
\arguments{
\item{nsid}{Netsuit ID}

\item{nspwd}{Netsuit password}
}
\value{
Three month CG storage fee over Profit
        One month CG storage fee over Profit
        Two files in N:/E Commerce/Public Share/Dot Com - Wayfair/CG Slow Item/
}
\description{
CG_Nmargin
}
\examples{

}
