% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewSalesWeekCG.R
\name{SaleWeekCG}
\alias{SaleWeekCG}
\title{NewSalesWeekCG}
\usage{
SaleWeekCG(nsid, nspwd, y)
}
\arguments{
\item{nsid}{Netsuit ID}

\item{nspwd}{Netsuit password}

\item{y}{yao or not}
}
\value{
a csv file in N:/E Commerce/Public Share/Dot Com - Promotion indicator/
}
\description{
NewSalesWeekCG
}
