% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFsales.R
\name{WFsales}
\alias{WFsales}
\title{WFsales}
\usage{
WFsales(nsid, nspwd, early, late, item)
}
\arguments{
\item{nsid}{}

\item{nspwd}{}

\item{early}{}

\item{late}{}

\item{item}{}
}
\value{
a data frame about the sales history
}
\description{
WFsales will help you grab the sales order under Wayfair's account. You can define the period and specific items you would like to know. The report name is WFinvoices which will tell you the margin, amount and the order name. The period is difined by "early" and "late", which means the query will return the perod from the day of system day - early to the day of system day - late.
}
\examples{
WFsales(nsid,nspwd,90,0, c("664WH","777[BED]"))

}
