% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeekOverWeekSales.R
\name{wow}
\alias{wow}
\title{WOW}
\usage{
wow(nsid, nspwd)
}
\arguments{
\item{nsid}{Netsuit ID}

\item{nspwd}{Netsuit password}
}
\description{
Week over week report evaluate how much sales order number and revenue different from last week and last year`s conterpart.
This report also compare the acumulated proformence with last year. All the result marked with differencial percentage by account.
}