% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewSalesWeekCG.R
\name{SaleWeekCG}
\alias{SaleWeekCG}
\title{Title NewSalesWeekCG}
\usage{
SaleWeekCG(nsid, nspwd, y)
}
\arguments{
\item{nsid}{Netsuit ID}

\item{nspwd}{Netsuit password}

\item{y}{yao or not}
}
\value{
a csv file in N:/E Commerce/Public Share/Dot Com - Promotion indicator/
}
\description{
This file is based on Jessy's original slow seller report. Will tell you howmany weeks the inventory can feed for each items.And will tell you the result for current inventory, the inventroy after 30 days and the inventory after 60 days. You can also find CG's feeding week.All the sales week are calculate base on either the recent one mont's sales record or 6 month's sales record.You can also find the sales speed and in-stock-rate in the output report.
}
